requirements.txt
scikit-image==0.17.2
scipy==1.7.3
pandas==1.2.1
numpy==1.19.5
matplotlib==3.3.3
opencv_python==4.5.1.48
imageio==2.9.0
Pillow==9.1.0
progressbar2==3.50.1